Make adjustments
